move file handling logic out of commands
